nand: Replace usage of loff_t with u32
The choice of loff_t as a type for an address is kind of peculiar, since
loff_t is a type "used for describing file sizes" that was introduced by
the Linux kernel.

Since all addresses used in the nand code seem to be 32-bit, and in
other parts of the code u32 is used to store addresses, we replace
loff_t with u32.

+++ nand/src/gpmc.cc
@@ -42,7 +42,7 @@ void Gpmc::add(Nand_chip *chip)
   _chips[0] = chip;
 }
 
-Nand_chip *Gpmc::select(loff_t /*addr*/)
+Nand_chip *Gpmc::select(u32 /*addr*/)
 {
   // XXX hack:we currently have only one chip
   return _chips[0];
